test_requirements.txt
black
flake8
isort
mypy
pandas
pytest
python-dotenv
tox
